+++ Sources/GrAIdient/Layer1D/BCE1D.swift
@@ -207,9 +207,9 @@ public class BCE1D: LayerOutput1D
     /// - Returns: The loss value.
     ///
     public func getLossGPU(
-        _ groundTruth: MetalBuffer<Float>,
+        _ groundTruth: MetalBuffer<Float16>,
         batchSize: Int,
-        nbNeurons: Int) throws -> Float
+        nbNeurons: Int) throws -> Float16
     {
         try checkGroundTruthGPU(
             groundTruth,
@@ -234,13 +234,13 @@ public class BCE1D: LayerOutput1D
         command.enqueue()
         
         MetalKernel.get.download([loss])
-        var loss: Float = 0.0
+        var loss: Float16 = 0.0
         let lossPtr = self.loss.buffer
         for i in 0..<batchSize
         {
             loss += lossPtr[i]
         }
-        return Float(coeff) * loss / Float(nbNeurons * batchSize)
+        return Float16(coeff) * loss / Float16(nbNeurons * batchSize)
     }
     
     ///
@@ -367,7 +367,7 @@ public class BCE1D: LayerOutput1D
     ///     - nbNeurons: Number of neurons.
     ///
     public func lossDerivativeGPU(
-        _ groundTruth: MetalBuffer<Float>,
+        _ groundTruth: MetalBuffer<Float16>,
         batchSize: Int,
         nbNeurons: Int) throws
     {
@@ -382,7 +382,7 @@ public class BCE1D: LayerOutput1D
             try layerPrev.checkStateBackwardGPU(batchSize: batchSize)
             
             let pNbNeurons: [UInt32] = [UInt32(nbNeurons)]
-            let pCoeff: [Float] = [Float(coeff)]
+            let pCoeff: [Float16] = [Float16(coeff)]
             let pNbBatch: [UInt32] = [UInt32(batchSize)]
             let pDirty: [UInt32] = layerPrev.dirty ? [1] : [0]
