+++ Sources/GrAIdient/Layer1D/Input1D.swift
@@ -63,7 +63,7 @@ class InputArrays1D: InputArrays<Layer1D>, IWeightArrays
 class InputBuffers1D: InputBuffers<Layer1D>, IWeightBuffers
 {
     /// Inputs buffer: the buffer to be update.
-    var w: MetalBuffer<Float>
+    var w: MetalBuffer<Float16>
     {
         get {
             return _layer.outs
@@ -71,7 +71,7 @@ class InputBuffers1D: InputBuffers<Layer1D>, IWeightBuffers
     }
     
     /// Gradients buffer.
-    var g: MetalBuffer<Float>
+    var g: MetalBuffer<Float16>
     {
         get {
             return _layer.delta
@@ -94,13 +94,13 @@ public class Input1D: LayerInput1D, LayerUpdate
     public var accumulateDeltaWeights: Bool = false
     
     /// Weights in the CPU execution context.
-    public var weightsCPU: [Float]
+    public var weightsCPU: [Float16]
     {
         get { [] }
         set {}
     }
     /// Weights in the GPU execution context.
-    public var weightsGPU: [Float]
+    public var weightsGPU: [Float16]
     {
         get { [] }
         set {}
@@ -304,7 +304,7 @@ public class Input1D: LayerInput1D, LayerUpdate
     ///     - nbNeurons: Number of neurons.
     ///
     public func setDataGPU(
-        _ data: MetalPrivateBuffer<Float>,
+        _ data: MetalPrivateBuffer<Float16>,
         batchSize: Int,
         nbNeurons: Int) throws
     {
